🔧 refactor: 공통 텍스트필드 겹치는 부분 분리

+++ finafanAndroid/app/src/main/java/com/a702/finafan/common/utils/StringUtil.kt
@@ -9,7 +9,7 @@ object StringUtil {
         return formatter.format(amount) + "원"
     }
 
-    fun formatEditMoney(amount: Int): String {
+    fun formatEditMoney(amount: Long): String {
         val formatter = NumberFormat.getNumberInstance(Locale.KOREA)
         return formatter.format(amount)
     }
